+++ src/modules/deeponet/training_strategies/helpers/__init__.py
@@ -1,5 +1,5 @@
 from .decomposition_helper import DecompositionHelper
-from .optimizer_manager import OptimizerSchedulerManager
+from ....pipe.optimizer_manager import OptimizerSchedulerManager
 from .pod_helper import PODBasisHelper
 from .phase_manager import PhaseManager
 from .two_step_helper import TwoStepHelper
